+++ src/ace_time_c/local_date.h
@@ -71,7 +71,7 @@ void atc_set_current_epoch_year(int16_t year);
  * The smallest year (inclusive) for which calculations involving the 32-bit
  * `epoch_seconds` and time zone transitions are guaranteed to be valid without
  * underflowing or overflowing. Valid years satisfy the condition `year >=
- * atc_local_valid_year_lower()`.
+ * atc_epoch_valid_year_lower()`.
  *
  * A 32-bit integer has a range of about 136 years, so the half interval is 68
  * years. But the algorithms to calculate transitions in `zone_processing.h` use
@@ -81,13 +81,13 @@ void atc_set_current_epoch_year(int16_t year);
  * smaller value in the future if the internal calculations can be verified to
  * avoid underflow or overflow problems.
  */
-int16_t atc_local_valid_year_lower();
+int16_t atc_epoch_valid_year_lower();
 
 /**
  * The largest year (exclusive) for which calculations involving the 32-bit
  * `epoch_seconds` and time zone transitions are guaranteed to be valid without
  * underflowing or overflowing. Valid years satisfy the condition `year <
- * atc_local_valid_year_upper()`.
+ * atc_epoch_valid_year_upper()`.
  *
  * A 32-bit integer has a range of about 136 years, so the half interval is 68
  * years. But the algorithms to calculate the transitions in `zone_processing.h`
@@ -97,7 +97,7 @@ int16_t atc_local_valid_year_lower();
  * larger value in the future if the internal calculations can be verified to
  * avoid underflow or overflow problems.
  */
-int16_t atc_local_valid_year_upper();
+int16_t atc_epoch_valid_year_upper();
 
 /** Return true if year is a leap year. */
 bool atc_is_leap_year(int16_t year);