+++ tests/local_date_test.c
@@ -185,8 +185,8 @@ ACU_TEST(test_epoch_year_and_valid_years)
   int16_t saved_epoch_year = atc_get_current_epoch_year();
 
   atc_set_current_epoch_year(2050);
-  ACU_ASSERT(2000 == atc_local_valid_year_lower());
-  ACU_ASSERT(2100 == atc_local_valid_year_upper());
+  ACU_ASSERT(2000 == atc_epoch_valid_year_lower());
+  ACU_ASSERT(2100 == atc_epoch_valid_year_upper());
 
   atc_set_current_epoch_year(saved_epoch_year);
 }